+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 渲染插件支持配置依赖项 | Render plugins support to configure dependencies.
 - 插件化渲染后处理方法 | Plugin rendering post-processing method.
 - 将插件对象添加至插件上下文 | Add the plugin object to the plugin context.
+- 新增命令行参数 `--workspace` 设置工作空间目录 | Add the command line argument `--workspace` to set the workspace directory.
 
 ## v0.1.2 / 2022-08-08